Improve SIGCHLD handling
Net::OpenSSH sets a custom dummy SIGCHLD handler in a couple of places
in order to ensure that syscalls are fail inmediately with EINTR when
a child process terminates.

This patch also removes an useless localization of $SIG{CHLD} and the
FAQ about setting $SIG{CHLD} before calling system & co that doesn't
apply anymore.