Force the use of nonblocking IO.
According to the rsync man page, it says that SSH *prefers* nonblocking IO;

--blocking-io
This tells rsync to use blocking I/O when launching a remote shell transport. If the remote shell is either rsh or remsh, rsync defaults to using blocking I/O, otherwise it defaults to using non-blocking I/O. (Note that ssh prefers non-blocking I/O.)

This fixes a bug we are seeing on AlmaLinux with stalled rsync transfers.